# Add Geometry/GeometryCollection forEachPoint/forEachGeometry