--- a/include/simfil/model/arena.h
+++ b/include/simfil/model/arena.h
@@ -190,6 +190,12 @@ public:
         friend class ArrayRange;
 
     public:
+        using iterator_category = std::input_iterator_tag;
+        using value_type = T;
+        using difference_type = std::ptrdiff_t;
+        using pointer = value_type*;
+        using reference = ElementRef;
+
         ArrayIterator(ArrayArenaRef arena, ArrayIndex array_index, size_t elem_index)
             : arena_(arena), array_index_(array_index), elem_index_(elem_index) {}
 
@@ -260,6 +266,12 @@ public:
         bool operator!=(const ArrayArenaIterator& other) const {
             return !(*this == other);  // NOLINT
         }
+
+        using iterator_category = std::input_iterator_tag;
+        using value_type = ArrayRange;
+        using difference_type = std::ptrdiff_t;
+        using pointer = value_type*;
+        using reference = value_type&;
 
     private:
         ArrayArena& arena_;