--- a/include/simfil/model/nodes.h
+++ b/include/simfil/model/nodes.h
@@ -56,7 +56,7 @@ using ScalarValueType = std::variant<
  * All ModelNode types are actually pointers into a ModelPool, via their
  * nested ModelNodeAddress. They keep the ModelPool which they reference
  * alive. The shared_model_ptr wrapper ensures that the user has
- * a better sense of the object they are dealing with. Only mapget::shared_model_ptr
+ * a better sense of the object they are dealing with. Only shared_model_ptr
  * is allowed to copy ModelNodes - so e.g. `auto node = *nodePtr` is not possible.
  */
 template<typename T>
@@ -293,7 +293,7 @@ protected:
     ScalarValueType data_;
 
     /// Reference to the model which controls this node
-    ModelConstPtr pool_;
+    ModelConstPtr model_;
 
     /// Address of the model pool node which this object references
     ModelNodeAddress addr_;
@@ -338,20 +338,26 @@ struct ValueNode final : public ModelNodeBase
 /**
  * ModelNode base class which provides a pool() method that returns
  * a reference to a Model-derived pool type.
- * @tparam PoolType Model-derived type.
+ * @tparam ModelType Model-derived type.
  */
-template<class PoolType>
-struct MandatoryDerivedModelPoolNodeBase : public ModelNodeBase
+template<class ModelType>
+struct MandatoryDerivedModelNodeBase : public ModelNodeBase
 {
 protected:
-    inline PoolType& pool() const {return *reinterpret_cast<PoolType*>(const_cast<Model*>(pool_.get()));}  // NOLINT
+    template<class ModelType_ = ModelType>
+    inline ModelType_* modelPtr() const {
+        static_assert(std::is_base_of<ModelType, ModelType_>::value);
+        return reinterpret_cast<ModelType_*>(const_cast<Model*>(model_.get()));
+    }  // NOLINT
 
-    MandatoryDerivedModelPoolNodeBase(ModelConstPtr p, ModelNodeAddress a={}, ScalarValueType data={})  // NOLINT
+    inline ModelType& model() const {return *modelPtr<ModelType>();}  // NOLINT
+
+    MandatoryDerivedModelNodeBase(ModelConstPtr p, ModelNodeAddress a={}, ScalarValueType data={})  // NOLINT
         : ModelNodeBase(p, a, std::move(data)) {}
-    MandatoryDerivedModelPoolNodeBase(ModelNode const& n) : ModelNodeBase(n) {}  // NOLINT
+    MandatoryDerivedModelNodeBase(ModelNode const& n) : ModelNodeBase(n) {}  // NOLINT
 };
 
-using MandatoryModelPoolNodeBase = MandatoryDerivedModelPoolNodeBase<ModelPool>;
+using MandatoryModelPoolNodeBase = MandatoryDerivedModelNodeBase<ModelPool>;
 
 namespace detail
 {
@@ -554,12 +560,15 @@ struct Geometry final : public MandatoryModelPoolNodeBase
      *  Must return true to continue iteration, false to abort iteration.
      * @return True if all points were visited, false if the callback ever returned false.
      * @example
-     *   collection->forEachPoint([](auto&& point){
+     *   collection->forEachPoint([](simfil::geo::Point<double>&& point){
      *      std::cout << point.x() << "," << point.y() << "," << point.z() << std::endl;
      *      return true;
      *   })
+     * @note The ModelType must also be templated here, because in this header
+     *  the class only exists in a predeclared form.
      */
-    bool forEachPoint(std::function<bool(geo::Point<double> const& point)> const& callback) const;
+    template <typename LambdaType, class ModelType=ModelPool>
+    bool forEachPoint(LambdaType const& callback) const;
 
 protected:
     struct Data {
@@ -611,12 +620,20 @@ struct GeometryCollection : public MandatoryModelPoolNodeBase
      *  Must return true to continue iteration, false to abort iteration.
      * @return True if all geometries were visited, false if the callback ever returned false.
      * @example
-     *   collection->forEachGeometry([](auto&& geom){
+     *   collection->forEachGeometry([](simfil::shared_model_ptr<Geometry> const& geom){
      *      std::cout << geom->type() << std::endl;
      *      return true;
      *   })
+     * @note The ModelType must also be templated here, because in this header
+     *  the class only exists in a predeclared form.
      */
-   bool forEachGeometry(std::function<bool(shared_model_ptr<Geometry> geom)> const& callback) const;
+    template <typename LambdaType, class ModelType=ModelPool>
+    bool forEachGeometry(LambdaType const& callback) const {
+        auto geomArray = modelPtr<ModelType>()->arrayMemberStorage().range((ArrayIndex)addr().index());
+        return std::all_of(geomArray.begin(), geomArray.end(), [this, &callback](auto&& geomNodeAddress){
+            return callback(modelPtr<ModelType>()->resolveGeometry(ModelNode::Ptr::make(model_, geomNodeAddress)));
+        });
+    }
 
 protected:
     using Storage = Array::Storage;
@@ -628,7 +645,7 @@ protected:
 struct VertexBufferNode final : public MandatoryModelPoolNodeBase
 {
     friend class ModelPool;
-    friend class Geometry;
+    friend struct Geometry;
 
     [[nodiscard]] ValueType type() const override;
     [[nodiscard]] ModelNode::Ptr at(int64_t) const override;
@@ -649,7 +666,7 @@ protected:
 struct VertexNode final : public MandatoryModelPoolNodeBase
 {
     friend class ModelPool;
-    friend class Geometry;
+    friend struct Geometry;
 
     [[nodiscard]] ValueType type() const override;
     [[nodiscard]] ModelNode::Ptr at(int64_t) const override;
@@ -664,5 +681,15 @@ protected:
     geo::Point<double> point_;
 };
 
+template <typename LambdaType, class ModelType>
+bool Geometry::forEachPoint(LambdaType const& callback) const {
+    VertexBufferNode vertexBufferNode{geomData_, model_, {ModelType::PointBuffers, addr_.index()}};
+    for (auto i = 0; i < vertexBufferNode.size(); ++i) {
+        VertexNode vertex{*vertexBufferNode.at(i), geomData_};
+        if (!callback(vertex.point_))
+            return false;
+    }
+    return true;
+}
 
 }