--- a/test/simfil.cpp
+++ b/test/simfil.cpp
@@ -375,6 +375,16 @@ TEST_CASE("Model Pool Validation", "[model.validation]") {
     pool->newObject()->addField("good", ModelNode::Ptr::make(pool, ModelNodeAddress{ModelPool::Objects, 666}));
     REQUIRE_THROWS(pool->validate());
 
+    // Recognize dangling array member pointer
+    pool->clear();
+    pool->newObject()->addField("good", ModelNode::Ptr::make(pool, ModelNodeAddress{ModelPool::Arrays, 666}));
+    REQUIRE_THROWS(pool->validate());
+
+    // Recognize dangling root
+    pool->clear();
+    pool->addRoot(ModelNode::Ptr::make(pool, ModelNodeAddress{ModelPool::Objects, 666}));
+    REQUIRE_THROWS(pool->validate());
+
     // An empty model should be valid
     pool->clear();
     REQUIRE_NOTHROW(pool->validate());