… test/ext-geo.cpp