I have added `Geometry::forEachPoint` and `GeometryCollection::forEachGeometry`. While doing this, I also noticed that the indirection from Object/Array ModelNode indices to ArrayIndices was completely redundant. Removing this indirection reduces the model size and brings small uplifts for all benchmarks.

**Before:**

```
32: Query typeof Recursive                         100             1     4.02004 s 
32:                                         38.5438 ms    38.4231 ms    38.6848 ms 
32:                                         661.456 us    543.953 us    822.811 us 
32:                                                                                
32: Query field id Recursive                       100             1     2.87265 s 
32:                                         28.6523 ms    28.5987 ms    28.6979 ms 
32:                                         251.096 us    213.424 us    304.459 us 
32:                                                                                
32: Query field id                                 100             1    53.2869 ms 
32:                                         527.767 us    523.889 us    537.298 us 
32:                                         28.7552 us     8.7424 us    50.7735 us 
```

**After:**

```
33: Query typeof Recursive                         100             1     3.57508 s 
33:                                         35.6623 ms    35.5108 ms    35.7912 ms 
33:                                          711.18 us    599.164 us    843.664 us 
33:                                                                                
33: Query field id Recursive                       100             1     2.78226 s 
33:                                         27.6927 ms    27.6588 ms    27.7258 ms 
33:                                          169.37 us    144.465 us    206.989 us 
33:                                                                                
33: Query field id                                 100             1    50.0698 ms 
33:                                         508.464 us    507.384 us    509.793 us 
33:                                         6.06923 us    5.04348 us    8.44147 us 
```